refactor(bonsai-archive): extract ArchiveTrieNodeEntry from TrieNodeDiffCodec
The decoded, typed view of a history entry (FULL/DIFF/tombstone) is its
own concept, not an implementation detail of the codec, and warrants a
clearer name than the nested Decoded class - both the metadata-byte
wire-format constants and the shape-specific accessors now live on
ArchiveTrieNodeEntry. TrieNodeDiffCodec keeps encode/decode/reconstruct.
